Add CMake subdirectory for benchmarks.
